[BBB-112] 🛠️Refactor: Feedback API Algo 도메인에서 Study 도메인으로 이동
