+++ app/external-api/src/main/java/com/bombombom/devs/external/algo/service/AlgorithmProblemService.java
@@ -3,27 +3,9 @@
 package com.bombombom.devs.external.algo.service;
 
 import com.bombombom.devs.algo.model.AlgorithmProblem;
-import com.bombombom.devs.algo.model.AlgorithmProblemFeedback;
-import com.bombombom.devs.algo.repository.AlgorithmProblemFeedbackRepository;
 import com.bombombom.devs.algo.repository.AlgorithmProblemRepository;
 import com.bombombom.devs.core.enums.AlgoTag;
-import com.bombombom.devs.core.exception.NotAcceptableException;
-import com.bombombom.devs.core.exception.NotFoundException;
-import com.bombombom.devs.core.util.Clock;
 import com.bombombom.devs.external.algo.config.ProbabilityConfig;
-import com.bombombom.devs.external.algo.service.dto.command.FeedbackAlgorithmProblemCommand;
-import com.bombombom.devs.study.model.AlgorithmProblemSolveHistory;
-import com.bombombom.devs.study.model.AlgorithmStudy;
-import com.bombombom.devs.study.model.Round;
-import com.bombombom.devs.study.model.Study;
-import com.bombombom.devs.study.model.StudyType;
-import com.bombombom.devs.study.repository.AlgorithmProblemAssignmentRepository;
-import com.bombombom.devs.study.repository.AlgorithmProblemSolveHistoryRepository;
-import com.bombombom.devs.study.repository.RoundRepository;
-import com.bombombom.devs.study.repository.StudyRepository;
-import com.bombombom.devs.study.repository.UserStudyRepository;
-import com.bombombom.devs.user.model.User;
-import com.bombombom.devs.user.repository.UserRepository;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -31,22 +13,13 @@ import java.util.Map;
 import java.util.random.RandomGenerator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
 @Service
 public class AlgorithmProblemService {
 
-    private final Clock clock;
     private final RandomGenerator randomGenerator;
     private final AlgorithmProblemRepository algorithmProblemRepository;
-    private final RoundRepository roundRepository;
-    private final StudyRepository studyRepository;
-    private final UserRepository userRepository;
-    private final UserStudyRepository userStudyRepository;
-    private final AlgorithmProblemAssignmentRepository algorithmProblemAssignmentRepository;
-    private final AlgorithmProblemFeedbackRepository algorithmProblemFeedbackRepository;
-    private final AlgorithmProblemSolveHistoryRepository algorithmProblemAssignmentSolveHistoryRepository;
 
 
     public void saveProblems(List<AlgorithmProblem> problems) {
@@ -96,121 +69,6 @@ public class AlgorithmProblemService {
             .orElse(null);
     }
 
-    @Transactional
-    public void feedback(Long userId,
-        FeedbackAlgorithmProblemCommand feedbackAlgorithmProblemCommand) {
 
-
-        Study study = studyRepository.findById(
-                feedbackAlgorithmProblemCommand.studyId())
-            .orElseThrow(() -> new NotFoundException("Study Not Found"));
-
-        if (study.getStudyType() != StudyType.ALGORITHM) {
-            throw new NotAcceptableException("Feedback can only be given to Algorithm Study");
-        }
-
-        AlgorithmStudy algorithmStudy = (AlgorithmStudy) study;
-
-        Round round = roundRepository.findRoundByStudyIdAndStartDateBeforeAndEndDateAfter(
-                algorithmStudy.getId(), clock.today())
-            .orElseThrow(() -> new NotFoundException("Ongoing Round Not Found"));
-
-        AlgorithmProblem problem = algorithmProblemRepository.findById(
-                feedbackAlgorithmProblemCommand.problemId())
-            .orElseThrow(() -> new NotFoundException("Problem Not Found"));
-
-
-        AlgorithmProblemSolveHistory history = algorithmProblemAssignmentSolveHistoryRepository.findByUserIdAndProblemId(userId, problem.getId())
-            .orElseThrow(() -> new NotFoundException("Solve History Not Found"));
-        
-        System.out.println("history.getSolvedAt() = " + history.getSolvedAt());
-        if( history.getSolvedAt() == null ){
-            throw new IllegalStateException("Cant Give Feedback On Unsolved Problem");
-        }
-
-
-
-
-        if(!algorithmProblemAssignmentRepository.existsByRoundIdAndProblemId(round.getId(),
-                problem.getId())){
-            throw new NotAcceptableException("Problem is not ongoing assignment");
-
-        }
-
-        if (!userStudyRepository.existsByUserIdAndStudyId(userId,
-            algorithmStudy.getId())) {
-            throw new NotAcceptableException("User is not a member");
-        }
-
-
-        User user = userRepository.findById(userId)
-            .orElseThrow(() -> new NotFoundException("User Not Found"));
-
-            AlgorithmProblemFeedback preFeedback =
-                algorithmProblemFeedbackRepository.findByUserIdAndProblemId(userId, problem.getId())
-                    .orElse(null);
-
-
-            AlgorithmProblemFeedback newFeedback = AlgorithmProblemFeedback.builder()
-                .problem(problem)
-                .user(user)
-                .again(feedbackAlgorithmProblemCommand.again())
-                .difficulty(feedbackAlgorithmProblemCommand.difficulty())
-                .build();
-
-            if(preFeedback != null){
-                changeFeedback(algorithmStudy, preFeedback, newFeedback);
-                preFeedback.update(newFeedback);
-
-            } else{
-                applyFeedback(algorithmStudy, newFeedback);
-                problem.addFeedback(newFeedback);
-                algorithmProblemRepository.save(problem);
-            }
-
-    }
-
-    private void applyFeedback(AlgorithmStudy study, AlgorithmProblemFeedback feedback) {
-        adjustDifficulty(study.getId(), feedback.getProblem().getTag(),
-            study.getDifficultyVariance(feedback));
-    }
-
-    private void changeFeedback(AlgorithmStudy study, AlgorithmProblemFeedback preFeedback,
-        AlgorithmProblemFeedback newFeedback) {
-        adjustDifficulty(study.getId(), preFeedback.getProblem().getTag(),
-            study.getDifficultyVariance(newFeedback) - study.getDifficultyVariance(preFeedback));
-    }
-
-    private void adjustDifficulty(Long studyId, AlgoTag tag, Float variance){
-
-        switch (tag) {
-            case AlgoTag.DP -> {
-                studyRepository.increaseDifficultyDpById(studyId,variance);
-            }
-            case AlgoTag.GEOMETRY -> {
-                studyRepository.increaseDifficultyGeometryById(studyId,variance);
-            }
-            case AlgoTag.DATA_STRUCTURES -> {
-                studyRepository.increaseDifficultyDataStructureById(studyId,variance);
-            }
-            case AlgoTag.GRAPHS -> {
-                studyRepository.increaseDifficultyGraphById(studyId,variance);
-            }
-            case AlgoTag.GREEDY -> {
-                studyRepository.increaseDifficultyGreedyById(studyId, variance);
-            }
-            case AlgoTag.IMPLEMENTATION -> {
-                studyRepository.increaseDifficultyImplementationById(studyId,variance);
-            }
-            case AlgoTag.MATH -> {
-                studyRepository.increaseDifficultyMathById(studyId,variance);
-            }
-            case AlgoTag.STRING -> {
-                studyRepository.increaseDifficultyStringById(studyId,variance);
-            }
-            default -> throw new IllegalStateException("Incorrect use of AlgoTag");
-        }
-
-    }
 
 }
