+++ app/external-api/src/main/java/com/bombombom/devs/external/study/service/StudyService.java
@@ -1,11 +1,16 @@
 package com.bombombom.devs.external.study.service;
 
 import com.bombombom.devs.algo.model.AlgorithmProblem;
+import com.bombombom.devs.algo.model.AlgorithmProblemFeedback;
+import com.bombombom.devs.algo.repository.AlgorithmProblemFeedbackRepository;
 import com.bombombom.devs.algo.repository.AlgorithmProblemRepository;
 import com.bombombom.devs.book.model.Book;
 import com.bombombom.devs.book.repository.BookRepository;
+import com.bombombom.devs.core.enums.AlgoTag;
+import com.bombombom.devs.core.exception.NotAcceptableException;
 import com.bombombom.devs.core.exception.NotFoundException;
 import com.bombombom.devs.core.util.Clock;
+import com.bombombom.devs.external.algo.service.dto.command.FeedbackAlgorithmProblemCommand;
 import com.bombombom.devs.external.study.service.dto.command.JoinStudyCommand;
 import com.bombombom.devs.external.study.service.dto.command.RegisterAlgorithmStudyCommand;
 import com.bombombom.devs.external.study.service.dto.command.RegisterBookStudyCommand;
@@ -15,12 +20,15 @@ import com.bombombom.devs.external.study.service.dto.result.StudyDetailsResult;
 import com.bombombom.devs.external.study.service.dto.result.StudyProgressResult;
 import com.bombombom.devs.external.study.service.dto.result.StudyResult;
 import com.bombombom.devs.external.study.service.factory.StudyServiceFactory;
-import com.bombombom.devs.job.AlgorithmProblemConverter;
+import com.bombombom.devs.study.model.AlgorithmProblemSolveHistory;
 import com.bombombom.devs.study.model.AlgorithmStudy;
 import com.bombombom.devs.study.model.BookStudy;
 import com.bombombom.devs.study.model.Round;
 import com.bombombom.devs.study.model.Study;
+import com.bombombom.devs.study.model.StudyType;
 import com.bombombom.devs.study.model.UserStudy;
+import com.bombombom.devs.study.repository.AlgorithmProblemAssignmentRepository;
+import com.bombombom.devs.study.repository.AlgorithmProblemSolveHistoryRepository;
 import com.bombombom.devs.study.repository.RoundRepository;
 import com.bombombom.devs.study.repository.StudyRepository;
 import com.bombombom.devs.study.repository.UserStudyRepository;
@@ -45,7 +53,9 @@ public class StudyService {
     private final RoundRepository roundRepository;
     private final StudyServiceFactory studyServiceFactory;
     private final AlgorithmProblemRepository algoProblemRepository;
-    private final AlgorithmProblemConverter algorithmProblemConverter;
+    private final AlgorithmProblemAssignmentRepository algoAssignmentRepository;
+    private final AlgorithmProblemSolveHistoryRepository algoSolveHistoryRepository;
+    private final AlgorithmProblemFeedbackRepository algoFeedbackRepository;
 
     @Transactional
     public AlgorithmStudyResult createAlgorithmStudy(
@@ -189,5 +199,116 @@ public class StudyService {
             .map(UserStudy::getUser).toList();
         return StudyProgressResult.fromEntity(study.getStudyType(), members,
             studyServiceFactory.getService(study.getStudyType()).findStudyProgress(round, members));
+    }
+
+
+    @Transactional
+    public void feedback(Long userId,
+        FeedbackAlgorithmProblemCommand feedbackAlgorithmProblemCommand) {
+
+        Study study = studyRepository.findById(
+                feedbackAlgorithmProblemCommand.studyId())
+            .orElseThrow(() -> new NotFoundException("Study Not Found"));
+
+        if (study.getStudyType() != StudyType.ALGORITHM) {
+            throw new NotAcceptableException("Feedback can only be given to Algorithm Study");
+        }
+
+        AlgorithmStudy algorithmStudy = (AlgorithmStudy) study;
+
+        Round round = roundRepository.findRoundByStudyIdAndStartDateBeforeAndEndDateAfter(
+                algorithmStudy.getId(), clock.today())
+            .orElseThrow(() -> new NotFoundException("Ongoing Round Not Found"));
+
+        AlgorithmProblem problem = algoProblemRepository.findById(
+                feedbackAlgorithmProblemCommand.problemId())
+            .orElseThrow(() -> new NotFoundException("Problem Not Found"));
+
+        AlgorithmProblemSolveHistory history = algoSolveHistoryRepository.findByUserIdAndProblemId(
+                userId, problem.getId())
+            .orElseThrow(() -> new NotFoundException("Solve History Not Found"));
+
+        if (history.getSolvedAt() == null) {
+            throw new IllegalStateException("Cant Give Feedback On Unsolved Problem");
+        }
+
+        if (!algoAssignmentRepository.existsByRoundIdAndProblemId(round.getId(),
+            problem.getId())) {
+            throw new NotAcceptableException("Problem is not ongoing assignment");
+
+        }
+
+        if (!userStudyRepository.existsByUserIdAndStudyId(userId,
+            algorithmStudy.getId())) {
+            throw new NotAcceptableException("User is not a member");
+        }
+
+        User user = userRepository.findById(userId)
+            .orElseThrow(() -> new NotFoundException("User Not Found"));
+
+        AlgorithmProblemFeedback preFeedback =
+            algoFeedbackRepository.findByUserIdAndProblemId(userId, problem.getId())
+                .orElse(null);
+
+        AlgorithmProblemFeedback newFeedback = AlgorithmProblemFeedback.builder()
+            .problem(problem)
+            .user(user)
+            .again(feedbackAlgorithmProblemCommand.again())
+            .difficulty(feedbackAlgorithmProblemCommand.difficulty())
+            .build();
+
+        if (preFeedback != null) {
+            changeFeedback(algorithmStudy, preFeedback, newFeedback);
+            preFeedback.update(newFeedback);
+
+        } else {
+            applyFeedback(algorithmStudy, newFeedback);
+            problem.addFeedback(newFeedback);
+            algoProblemRepository.save(problem);
+        }
+
+    }
+
+    private void applyFeedback(AlgorithmStudy study, AlgorithmProblemFeedback feedback) {
+        adjustDifficulty(study.getId(), feedback.getProblem().getTag(),
+            study.getDifficultyVariance(feedback));
+    }
+
+    private void changeFeedback(AlgorithmStudy study, AlgorithmProblemFeedback preFeedback,
+        AlgorithmProblemFeedback newFeedback) {
+        adjustDifficulty(study.getId(), preFeedback.getProblem().getTag(),
+            study.getDifficultyVariance(newFeedback) - study.getDifficultyVariance(preFeedback));
+    }
+
+    private void adjustDifficulty(Long studyId, AlgoTag tag, Float variance) {
+
+        switch (tag) {
+            case AlgoTag.DP -> {
+                studyRepository.increaseDifficultyDpById(studyId, variance);
+            }
+            case AlgoTag.GEOMETRY -> {
+                studyRepository.increaseDifficultyGeometryById(studyId, variance);
+            }
+            case AlgoTag.DATA_STRUCTURES -> {
+                studyRepository.increaseDifficultyDataStructureById(studyId, variance);
+            }
+            case AlgoTag.GRAPHS -> {
+                studyRepository.increaseDifficultyGraphById(studyId, variance);
+            }
+            case AlgoTag.GREEDY -> {
+                studyRepository.increaseDifficultyGreedyById(studyId, variance);
+            }
+            case AlgoTag.IMPLEMENTATION -> {
+                studyRepository.increaseDifficultyImplementationById(studyId, variance);
+            }
+            case AlgoTag.MATH -> {
+                studyRepository.increaseDifficultyMathById(studyId, variance);
+            }
+            case AlgoTag.STRING -> {
+                studyRepository.increaseDifficultyStringById(studyId, variance);
+            }
+            default -> throw new IllegalStateException("Incorrect use of AlgoTag");
+        }
+
     }
 }
